feat: persist billable metrics
Currently, Orb API errors when using a cursor or a limit closer to 100, that should be fixed soon though on Orb's end and we can work around it by using the createdAt parameter when syncing.

There is also no billable_metric.created event and the updated event contains no id, so we are re-fetching all billable metrics right now.